renames api-key.ts to apiKey.ts
